Added a bunch of css
