+++ src/components/01-main-sections/01-intro-section/IntroSection.jsx
@@ -1,0 +1,17 @@
+import { TitleSimonFrisk } from "../../sections/01-intro-section/TitleSimonFrisk"
+import { TileImages } from "../../sections/01-intro-section/TileImages"
+import { WhoAmI } from "../../sections/01-intro-section/WhoAmI"
+import "./introSection.css"
+
+export const IntroSection = () => {
+  return (
+    <>
+      <section className="intro-section">
+        <TitleSimonFrisk />
+        <TileImages />
+        <WhoAmI />
+      </section>
+    </>
+
+  )
+}
