+++ src/components/01-main-sections/01-intro-section/introSection.css
@@ -1,0 +1,32 @@
+.intro-section {
+  /* height: 100dvh; */
+}
+
+.hello {
+  position: relative;
+  /* background-color: red; */
+  height: 250px;
+  width: 100%;
+}
+
+.intro-img-1 {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(0%, -45%) rotate(5deg);
+}
+
+.intro-img-2 {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  z-index: 2;
+}
+
+.intro-img-3 {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-100%, -45%) rotate(-5deg);
+}
