--- src/components/01-main-sections/01IntroSection.jsx
@@ -1,16 +1,0 @@
-import { TitleSimonFrisk } from "../sections/01-intro-section/TitleSimonFrisk"
-import { TileImages } from "../sections/01-intro-section/TileImages"
-import { WhoAmI } from "../sections/01-intro-section/WhoAmI"
-
-export const IntroSection = () => {
-  return (
-    <>
-      <section>
-        <TitleSimonFrisk />
-        <TileImages />
-        <WhoAmI />
-      </section>
-    </>
-
-  )
-}
