+++ src/components/01-main-sections/02-tech-section/TechSection.jsx
@@ -1,8 +1,9 @@
+import "./tech-section.css"
 
 export const TechSection = () => {
   return (
     <>
-      <section className="tech-section">
+      <section className="black-section">
         <h2>Tech</h2>
         <p>HTML, CSS, JavaScript, ES6, TypeScript, React, React Hooks, Web Accessibility, APIs, mob-programming, pair-programming, GitHub. </p>
       </section>
