+++ src/components/01-main-sections/03-featured-project-section/FeaturedProjectsSection.jsx
@@ -1,5 +1,5 @@
-import { ProjectCard } from "../sections/03-featured-project-section/ProjectCard"
-import { CardData } from "../sections/03-featured-project-section/CardData"
+import { ProjectCard } from "../../sections/03-featured-project-section/ProjectCard"
+import { CardData } from "../../sections/03-featured-project-section/CardData"
 
 
 export const FeaturedProjectSection = () => {
