+++ src/components/01-main-sections/04-skills-section/skills-section.css
@@ -1,0 +1,19 @@
+.skills {
+  display: flex;
+  flex-direction: column;
+}
+
+.skill-title {
+  border: solid 2px rgb(255, 255, 255);
+  padding: 6px;
+  border-radius: 4px;
+}
+
+@media (min-width: 1024px) {
+  .skills {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    column-gap: 24px;
+  }
+}
