+++ src/components/01-main-sections/05-my-words/my-words.css
@@ -1,0 +1,12 @@
+.word-card {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: left;
+}
+
+.date {
+  border: solid black 1px;
+  padding: 6px;
+  border-radius: 4px;
+}
