+++ src/components/01-main-sections/06-contact/Contact.jsx
@@ -2,9 +2,9 @@
 export const Contact = () => {
   return (
     <>
-      <section>
+      <section className="black-section">
         <h2>Lets talk</h2>
-        <img src="" alt="" />
+        <img src="/simon5.webp" alt="" />
         <p>Simon Frisk</p>
         <p>+46(0)73 433 30 61</p>
         <p>simoncarlfrisk@gmail.com</p>
