+++ src/components/sections/04-skills-section/SkillCategory.jsx
@@ -2,7 +2,7 @@
 export const SkillCategory = ({ title, skills }) => {
   return (
     <div>
-      <h3>{title}</h3>
+      <h3 className="skill-title">{title}</h3>
       {skills.map((skill, index) => (
         <p key={index}>{skill}</p>
       ))}
