+++ src/index.css
@@ -45,36 +45,7 @@ section {
   padding: 64px 50px;
 }
 
-.hello {
-  position: relative;
-  /* background-color: red; */
-  height: 250px;
-  width: 100%;
-}
-
-.intro-img-1 {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(0%, -45%) rotate(5deg);
-}
-
-.intro-img-2 {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  z-index: 2;
-}
-
-.intro-img-3 {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-100%, -45%) rotate(-5deg);
-}
-
-.tech-section {
+.black-section {
   background-color: black;
   color: white;
 }
